Updated to use of double instead of integer

+++ getAvg.c
@@ -1,6 +1,6 @@
-float SL_getAvg(int eingabe){
-	static int sum, n;
+double SL_getAvg(double eingabe){
+	static double sum, n;
 	sum += eingabe;
-	return(((float)sum)/ ++n);
+	return sum/++n;
 }
