+++ getStddev.c
@@ -1,5 +1,5 @@
-float SL_getStddev(int eingabe, float avg){
-	static int quadr_sum, n;
+double SL_getStddev(double eingabe, double avg){
+	static double quadr_sum, n;
 	quadr_sum += (eingabe - avg)*(eingabe - avg);
 	return quadr_sum;
 }
